Update DQMGUI Dockerfile to version HG1510f
The DQMGUI Dockerfile has been reviewed and updated to bring the most
up-to-date version of the software.

Anchillary packages have been added on top of the cms-dqmgui package in order
to be able to locally build the server, since they are at present required only
at build-time and not listed as runtime dependencies.

With the current docker image the user is able to immediately run the server
out of the box and also to develop it: for this last step we recommend to run
the docker container mounting the proper DQMGUI source repository from an
external volume.

The application must be installed with a user different from root, for security
reasons built inside cms software stack.  The application is run under the root
account to ease the process of file creation/deletion on an externally mounted
volumes.